Revertido a estado base. Eliminado menu de actualizacion

+++ GestorPreciosBarcode/Globales.vb
@@ -1,6 +1,6 @@
 ﻿Public Class Globales
     Public Shared libro As New List(Of Libros)
-    Public Shared tasa As New List(Of Single)
+    Public Shared tasa As New List(Of Integer)
     Public Shared historial As New List(Of EntradaHistorial)
     Public Shared hora As String
 End Class
